src/icland/world_gen/requirements.txt
numpy
hilbertcurve>=2
imageio
matplotlib
scipy
numpy-stl

# Testing
pytest
types-setuptools

# Documentation.
sphinx
